files structure changed, minor bug fixed
