+++ AniRPG.Application/Game/CharacterSystem/Constants/CharacterConstants.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace AniRPG.Application.Game.MapSystem.Constants
+namespace AniRPG.Application.Game.CharacterSystem.Constants
 {
     public static class CharacterConstants
     {
