+++ AniRPG.Application/Game/MapSystem/UseCases/Locations/Commands/UpdateLocation/UpdateLocationCommandHandler.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using MediatR;
 using AniRPG.Application.Game.MapSystem.Repositories;
-using AniRPG.Application.Game.MapSystem.Exceptions;
+using AniRPG.Application.Game.Common.Exceptions;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -31,6 +31,8 @@ namespace AniRPG.Application.Game.MapSystem.UseCases.Locations.Commands.UpdateLo
             location.Transitions.Clear();
             foreach (var transition in request.Transitions)
                 location.Transitions.Add(transition);
+
+            await _locationRepository.UpdateLocation(location);
         }
     }
 }
