+++ AniRPG.Application/Repositories/ICharacterRepository.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using AniRPG.Application.Game.CharacterSystem.Repositories;
+
+namespace AniRPG.Application.Repositories
+{
+    public interface ICharacterRepository : ICharacterSystemCharacterRepository
+    {
+
+    }
+}
