+++ AniRPG.Application/Repositories/ILocationRepository.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using AniRPG.Application.Game.MapSystem.Repositories;
+
+namespace AniRPG.Application.Repositories
+{
+    public interface ILocationRepository : IMapSystemLocationRepository
+    {
+
+    }
+}
